// ap-invoice-intern-feature-ap/frontend/types/api_schemas.ts
export interface AuditLogResponse {
  invoice_id?: number | null;
  action: string;
  status_before?: string | null;
  status_after?: string | null;
  performed_by?: string;
  details?: any | null;
  id: number;
  timestamp: string;
}